Make AnchorButton look like a dropdown anchor
Shows a down arrow at the right of the button label
